git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20240111_nbraddy_q13124@23499 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ pages/download.php
@@ -390,7 +390,7 @@ if(isset($filename))
     }
 else
     {   
-    header('Content-Disposition: inline;');
+    header("Content-Disposition: inline; filename=\"download.{$ext}\"");
     header('Content-Transfer-Encoding: binary');
     debug("PAGES/DOWNLOAD.PHP: Set header for streamed file");
     }
@@ -491,7 +491,7 @@ if(!hook('replacefileoutput'))
 
 // Deleting Exiftool temp File:
 // Note: Only for downloads (not previews)
-if(!$noattach && -1 == $alternative && $exiftool_write && file_exists($tmpfile))
+if(!$noattach && -1 == $alternative && $exiftool_write && isset($tmpfile) && file_exists($tmpfile))
     {
     delete_exif_tmpfile($tmpfile);
     }
